refactor(models): move BotModeratorRole into models/admin and expose checks publicly
- Rename models/botmod.py → models/admin.py to co-locate the BotModeratorRole
  model with the admin module that owns it
- Expose _reject → reject and _is_bot_moderator → is_bot_moderator as public
  API so callers (other modules, tests) can import them without the leading
  underscore convention implying they're private implementation details
- Update all import sites in admin.py, checks.py, and test_checks.py
- Refresh docs/admin.md to reflect the new internal symbol names
